# Add schools import script