--- a/Gemfile
+++ b/Gemfile
@@ -37,6 +37,7 @@ gem "propshaft"
 gem "puma", "~> 6.4"
 gem "pundit"
 gem "rainbow"
+gem "rubyzip"
 gem "sentry-rails"
 gem "sentry-ruby"
 gem "silencer", require: false