@tvararu 2024-08-08
```bash
$ heroku run bash -a mavis-pr-1586
Running bash on ⬢ mavis-pr-1586... up, run.3671 (Basic)
~ $ bin/rails console
Loading production environment (Rails 7.1.3.4)
irb(main):001> Location.count
=> 51128
```