@tvararu 2024-08-13
I attempted to write a download script, and got this far, but it fails and I'm not sure it's worth using. Leaving here for posterity:

```ruby
  # To get the latest version of the zip:
  # 1. Go to https://get-information-schools.service.gov.uk/Downloads
  # 2. Check "Establishment fields CSV"
  # 3. Submit
  # 4. Download the zip file
  # 5. Place it in db/data/edubasealldata.zip
  #
  # Alternatively, you can run this task.
  desc "Download schools data"
  task download: :environment do
    require "capybara/cuprite"

    puts "Starting schools data download process..."
    puts "Visiting the downloads page"
    browser = Capybara::Session.new(:cuprite)
    browser.visit "https://get-information-schools.service.gov.uk/Downloads"

    puts "Checking the establishment fields CSV checkbox"
    browser.find("#establishment-fields-csv-checkbox", visible: false).check

    puts "Clicking the submit button"
    browser.click_button "Download selected files"

    puts "Waiting for the 'Results.zip' button to appear..."
    browser.find_button("Results.zip", wait: 20)
    puts "'Results.zip' button found"

    puts "Clicking the 'Results.zip' button to download the file..."
    browser.click_button("Results.zip")

    puts "Waiting for the file to download..."
    # Wait for the file to be downloaded (adjust the timeout as needed)
    downloaded_file =
      Dir.glob("#{Capybara.save_path}/*.zip").max_by { |f| File.mtime(f) }
    wait_time = 0
    until downloaded_file && File.size(downloaded_file).positive?
      sleep 1
      wait_time += 1
      break if wait_time > 60 # Timeout after 60 seconds
      downloaded_file =
        Dir.glob("#{Capybara.save_path}/*.zip").max_by { |f| File.mtime(f) }
    end

    if downloaded_file
      puts "File downloaded successfully"
      FileUtils.mv(downloaded_file, "db/data/edubasealldata.zip")
      puts "File moved to db/data/edubasealldata.zip"
    else
      puts "Failed to download the file"
    end

    puts "Schools data downloaded to db/data/edubasealldata.zip"
    puts "Download process completed"
  end
```