@misaka 2024-08-13
I had a look at the download thing, it looks like Capybara wasn't saving the file anywhere? Mechanize seems to go a better job, I added `mechanize` to `Rubygems` and then converted that script to use it:

```ruby
# To get the latest version of the zip:
# 1. Go to https://get-information-schools.service.gov.uk/Downloads
# 2. Check "Establishment fields CSV"
# 3. Submit
# 4. Download the zip file
# 5. Place it in db/data/edubasealldata.zip
#
# Alternatively, you can run this task.
desc "Download schools data"
task "gias:download": :environment do
  require "mechanize"

  puts "Starting schools data download process..."
  agent = Mechanize.new

  puts "Visiting the downloads page"
  page = agent.get("https://get-information-schools.service.gov.uk/Downloads")

  puts "Checking the establishment fields CSV checkbox"
  form = page.form_with(action: "/Downloads/Collate")
  form.checkbox_with(id: "establishment-fields-csv-checkbox").check

  puts "Submitting the form"
  download_page = form.submit

  # There is a meta refresh on the download_page. Mechanize didn't seem to
  # follow it so we're just refreshing that page manually until the button shows
  # up below.
  wait_time = 0
  until (
          download_form =
            download_page.form_with(action: "/Downloads/Download/Extract")
        ) || wait_time > 60
    puts "Waiting for the 'Results.zip' link to appear..."
    sleep(2)
    wait_time += 2
    download_page_uri = download_page.uri
    download_page = agent.get(download_page_uri)
  end

  if download_form
    download_button = download_form.button_with(value: "Results.zip")
    puts "'Results.zip' link found, downloading the file..."
    download_file = agent.click(download_button)
    download_file.save
    puts "File downloaded successfully to #{download_file.filename}"
  else
    puts "Download button never appeared, abotting"
  end
end
```

And it seems to work:

```shell
$ rails gias:download
Starting schools data download process...
Visiting the downloads page
Checking the establishment fields CSV checkbox
Submitting the form
Waiting for the 'Results.zip' link to appear...
Waiting for the 'Results.zip' link to appear...
'Results.zip' link found, downloading the file...
File downloaded successfully to extract.zip
$ unzip -v extract.zip
Archive:  extract.zip
 Length   Method    Size  Cmpr    Date    Time   CRC-32   Name
--------  ------  ------- ---- ---------- ----- --------  ----
64588771  Defl:N  8004870  88% 08-13-2024 13:07 a2311d70  edubasealldata20240813.csv
--------          -------  ---                            -------
64588771          8004870  88%                            1 file
```